[FINNA-4067] Configurable header bar, footer menu and site map
