+++ local/config/finna/HeaderBar.yaml.sample
@@ -1,0 +1,70 @@
+# This file configures the HeaderBar section plugin, which renders menus and other
+# controls at the top of every page. If this file is empty or missing, default
+# settings will be used, matching the examples provided below.
+#
+# The format of this configuration is as follows:
+#
+# Array keys for every group are:
+#   - MenuItems: items shown in the menu - required
+#   - checkMethod: the name of a HeaderBar section plugin method to perform
+#     a check whether to show the group. If omitted, group will always display.
+#   - wrapper: setting a value for this key adds a wrapper <div> for the group.
+#     use attributes subkey for additional attributes for the wrapper.
+#   - attributes: additional attributes for the group's <ul> tag
+#   - navibarIni: setting this to true adds menu items configured in navibar.ini.
+#     These items will be the first items in the group before any additional items
+#     specified in MenuItems.
+#
+# Array keys for every menu item could be:
+#   - checkMethod: the name of a HeaderBar section plugin method to perform
+#     a check whether to show the item. If omitted, item will always display.
+#   - template: template to render the item with. Templates do not support the keys
+#     documented below.
+#   - name: name of an item, used as a CSS class for the menu item's <li> tag
+#   - icon: icon used for the item, has to be defined in theme config
+#   - attributes: additional attributes for the menu item's <a> tag
+#   - submenuItems: items to be shown in a dropdown menu. Submenu items support only
+#     the keys documented below.
+#   - label: the text shown as link, will be translated - required
+#   - route: route name used to generate link target - required unless url, template
+#     or submenuItems is set
+#   - routeParams: parameters for dynamic routes
+#   - url: target url, alternative to route - required unless route, template or
+#     submenuItems is set
+#
+# You can use checkMethod: alwaysFail to disable specific inherited groups or items
+# in local configuration. If you wish to completely override the default
+# configuration you can disable configuration inheritance by adding a
+# "@parent_yaml": false entry at the very top of your local configuration file.
+
+"@parent_yaml": false
+
+Header:
+  wrapper:
+    attributes:
+      id: header-collapse
+      class:
+        - collapse
+        - navbar-collapse
+  attributes:
+    class: navbar-main
+  navibarIni: true
+  MenuItems:
+    - template: Section/HeaderBar/HeaderBar-feedback.phtml
+      checkMethod: checkFeedback
+
+    - template: Section/HeaderBar/HeaderBar-cart.phtml
+      checkMethod: checkCart
+
+    - template: Section/HeaderBar/HeaderBar-themeOptions.phtml
+      checkMethod: checkThemeOptions
+
+HeaderRight:
+  attributes:
+    class: header-navbar-right
+  MenuItems:
+    - template: Section/HeaderBar/HeaderBar-allLangs.phtml
+      checkMethod: checkAllLangs
+
+    - template: Section/HeaderBar/HeaderBar-account.phtml
+      checkMethod: checkAccount
