+++ themes/finna2/templates/Section/HeaderBar/HeaderBar-feedback.phtml
@@ -1,0 +1,19 @@
+<!-- START of: finna - header-feedback.phtml -->
+<?php
+  $url = $this->url('feedback-home');
+  $linkAttrs = $this->htmlAttributes(
+      [
+          'href' => $url,
+          'class' => ['nav-link'],
+          'id' => 'feedback-link',
+      ]
+  );
+  if ($this->sectionPlugin->isCurrentPage($url)) {
+    $linkAttrs->add('class', 'active');
+    $linkAttrs->add('aria-current', 'page');
+  }
+?>
+<li class="nav-item" role="none">
+  <a<?=$linkAttrs?>><?=$this->transEsc('Feedback')?></a>
+</li>
+<!-- END of: finna - header-feedback.phtml -->
